tests/requirements.txt
mock ; python_version <= '3.3'
pytest-django
six
typing ; python_version <= '3.4'
